+++ flutter_theoplayer_sdk/flutter_theoplayer_sdk_platform_interface/lib/pigeon/apis.g.dart
@@ -603,6 +603,9 @@ abstract class THEOplayerFlutterTextTracksAPI {
   void onTextTrackAddDateRangeCue(int textTrackUid, String id, int uid, double startTime, double endTime, String? cueClass, double startDateMillis, double? endDateMillis, double? duration,
       double? plannedDuration, bool endOnNext, String? customAttributesJson, Uint8List? scte35Cmd, Uint8List? scte35Out, Uint8List? scte35In);
 
+  void onTextTrackUpdateDateRangeCue(int textTrackUid, String id, int uid, double startTime, double endTime, String? cueClass, double startDateMillis, double? endDateMillis, double? duration,
+      double? plannedDuration, bool endOnNext, String? customAttributesJson, Uint8List? scte35Cmd, Uint8List? scte35Out, Uint8List? scte35In);
+
   void onTextTrackRemoveCue(int textTrackUid, int cueUid);
 
   void onTextTrackEnterCue(int textTrackUid, int cueUid);
@@ -780,6 +783,53 @@ abstract class THEOplayerFlutterTextTracksAPI {
           try {
             api.onTextTrackAddDateRangeCue(arg_textTrackUid!, arg_id!, arg_uid!, arg_startTime!, arg_endTime!, arg_cueClass, arg_startDateMillis!, arg_endDateMillis, arg_duration, arg_plannedDuration,
                 arg_endOnNext!, arg_customAttributesJson, arg_scte35Cmd, arg_scte35Out, arg_scte35In);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          } catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
+        });
+      }
+    }
+    {
+      final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
+      if (api == null) {
+        pigeonVar_channel.setMessageHandler(null);
+      } else {
+        pigeonVar_channel.setMessageHandler((Object? message) async {
+          assert(message != null, 'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null.');
+          final List<Object?> args = (message as List<Object?>?)!;
+          final int? arg_textTrackUid = (args[0] as int?);
+          assert(
+              arg_textTrackUid != null, 'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null, expected non-null int.');
+          final String? arg_id = (args[1] as String?);
+          assert(arg_id != null, 'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null, expected non-null String.');
+          final int? arg_uid = (args[2] as int?);
+          assert(arg_uid != null, 'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null, expected non-null int.');
+          final double? arg_startTime = (args[3] as double?);
+          assert(
+              arg_startTime != null, 'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null, expected non-null double.');
+          final double? arg_endTime = (args[4] as double?);
+          assert(arg_endTime != null, 'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null, expected non-null double.');
+          final String? arg_cueClass = (args[5] as String?);
+          final double? arg_startDateMillis = (args[6] as double?);
+          assert(arg_startDateMillis != null,
+              'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null, expected non-null double.');
+          final double? arg_endDateMillis = (args[7] as double?);
+          final double? arg_duration = (args[8] as double?);
+          final double? arg_plannedDuration = (args[9] as double?);
+          final bool? arg_endOnNext = (args[10] as bool?);
+          assert(arg_endOnNext != null, 'Argument for dev.flutter.pigeon.theoplayer_platform_interface.THEOplayerFlutterTextTracksAPI.onTextTrackUpdateDateRangeCue was null, expected non-null bool.');
+          final String? arg_customAttributesJson = (args[11] as String?);
+          final Uint8List? arg_scte35Cmd = (args[12] as Uint8List?);
+          final Uint8List? arg_scte35Out = (args[13] as Uint8List?);
+          final Uint8List? arg_scte35In = (args[14] as Uint8List?);
+          try {
+            api.onTextTrackUpdateDateRangeCue(arg_textTrackUid!, arg_id!, arg_uid!, arg_startTime!, arg_endTime!, arg_cueClass, arg_startDateMillis!, arg_endDateMillis, arg_duration,
+                arg_plannedDuration, arg_endOnNext!, arg_customAttributesJson, arg_scte35Cmd, arg_scte35Out, arg_scte35In);
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);